Fundamental Fix of Instertion and Deletion Probabilities
Fixed a large bug that caused insertion and deletion probabilities to be set to 0

Interesting behavior. Up until NOR we can evolve fine but the NOR - XOR step proves to be very difficult. Instead of evolving XOR we obtained very short organisms that can still replicate despite a very high insertion probability.

Want to inspect the origin of random organisms of length 10, which is also the minimum allowed length,